Modernize alloc.c and subr.c

+++ ancient-src/v6/filsys.h
@@ -11,12 +11,12 @@
  */
 struct	filsys
 {
-	int	s_isize;	/* size in blocks of I list */
-	int	s_fsize;	/* size in blocks of entire volume */
-	int	s_nfree;	/* number of in core free blocks (0-100) */
-	int	s_free[100];	/* in core free blocks */
-	int	s_ninode;	/* number of in core I nodes (0-100) */
-	int	s_inode[100];	/* in core free I nodes */
+	int16_t	s_isize;	/* size in blocks of I list */
+	int16_t	s_fsize;	/* size in blocks of entire volume */
+	int16_t	s_nfree;	/* number of in core free blocks (0-100) */
+	int16_t	s_free[100];	/* in core free blocks */
+	int16_t	s_ninode;	/* number of in core I nodes (0-100) */
+	int16_t	s_inode[100];	/* in core free I nodes */
 	char	s_flock;	/* lock during free list manipulation */
 	char	s_ilock;	/* lock during I list manipulation */
 	char	s_fmod;		/* super block modified flag */
